backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
python-multipart==0.0.6
pymupdf==1.23.6
python-docx==1.1.0
pydantic==2.5.0
scikit-learn==1.3.2
spacy==3.7.2
nltk==3.8.1
pandas==2.1.3
numpy==1.26.2